refactor: collapse repl and server module split into app module
