+++ src/app.module.ts
@@ -1,15 +1,18 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
+import { ScheduleModule } from '@nestjs/schedule';
 import { NecordPaginationModule } from '@necord/pagination';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { NecordModule } from 'necord';
 import { IntentsBitField } from 'discord.js';
 import { InventoryModule } from './inventory/inventory.module';
 import { DiscordModule } from './discord/discord.module';
+import { EventsModule } from './events/events.module';
 import { GameModule } from './game/game.module';
 import { PermissionsService } from './permissions.service';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { validationSchema } from './app.config';
 import { MigrationKillSwitch as MigrationService } from './migrations.service';
 import * as migrations from './database/migrations';
 
@@ -62,4 +65,35 @@ import * as migrations from './database/migrations';
   providers: [MigrationService, PermissionsService, AppService],
   exports: [],
 })
-export class AppModule {}
+export class AppModule {
+  static getAppModules() {
+    return [
+      ConfigModule.forRoot({
+        envFilePath: [
+          '.env.local',
+          `.env.${process.env.NODE_ENV}.local`,
+          `.env.${process.env.NODE_ENV}`,
+          '.env',
+        ],
+        isGlobal: true,
+        cache: true,
+        validationSchema,
+      }),
+      this,
+    ];
+  }
+
+  static getServerModules() {
+    return [...this.getAppModules(), ScheduleModule.forRoot(), EventsModule];
+  }
+}
+
+@Module({
+  imports: AppModule.getServerModules(),
+})
+export class ServerModule {}
+
+@Module({
+  imports: AppModule.getAppModules(),
+})
+export class ReplModule {}
