+++ src/main.ts
@@ -1,6 +1,6 @@
 import { NestFactory } from '@nestjs/core';
 import { ConfigService } from '@nestjs/config';
-import { ServerModule } from './server.module';
+import { ServerModule } from './app.module';
 
 async function bootstrap() {
   const app = await NestFactory.create(ServerModule, {
