+++ src/repl.ts
@@ -1,5 +1,5 @@
 import { repl } from '@nestjs/core';
-import { ReplModule } from './repl.module';
+import { ReplModule } from './app.module';
 
 async function bootstrap() {
   await repl(ReplModule);
